[29158] Allow lazily loading a module and component
This fixes #29158, but is quite a hack:

1. We load the lazy calendar module

2. We get the lazy loadable component and put it into a
ngComponentOutlet. `ng-dynamic-component` will not work in this cases
because it uses a `ComponentFactoryResolver` from the including module,
not from the included module and thus the dynamic component is not known
(or rather, not in the `entryComponents` set). It would work if it were
to use the injector to get the CFR.

3. We then render another component that is able to set inputs due to 2.

https://community.openproject.com/wp/29158